Fix wpr recording
If try to record WPR for benchmark 'system_health.common_mobile' you
would get an AttributeError, which says that class MobileCommonSystemHealth
does not possess attribute 'allow_software_compositing'. It happens,
beacause this attribute is used in benchmark's method 'SetExtraBrowserOptions',
while this attribute is set in benchmark's method 'ProcessCommandLineArgs'.
The main problem lives in class 'WprRecorder' in record_wpr.py: benchmark's method
'SetExtraBrowserOptions' is called before 'ProcessCommandLineArgs', so
class variable 'allow_software_compositing' is not set when python tries
to use it.
The main purpose of this CL is to repair this bug by calling
'ProcessCommandLineArgs' before 'SetExtraBrowserOptions' in record_wpr.py

Change-Id: If68f1a57475127d1eec0af641ecdc118e5db10ba