startup.mobile: introduce experimental_network_request_start_time
On a trace downloaded from chromeperf I can see that this new metric
reveals that the request reaches /net with a delay of 86ms from the
start of navigation. We should rather target this metric in our
optimizations instead of navigation start.

Prefixing with 'experimental_' to check whether it works as expected on
all bots and local startup scenarios.

Bug: chromium:1302992
Change-Id: If82441356ea3079f7cc44954d8251762b26ac84f