[NetLog Viewer] Support new location of disabled DoH providers
We are moving the list of disabled DoH providers out of the DnsConfig
[0]. As of CL 3526375, it is emitted by GetNetInfo() [1]. This change
adds support for the new location and continues implicitly supporting
the old format, since DnsView just iterates over the DnsConfig fields.

This change also fixes a bug I ran into in DnsView where
onHostResolverInfoChanged() tried to render the addressess associated
with a cache entry. Not all entries are guaranteed to have addresses!

[0]: https://chromium-review.googlesource.com/c/chromium/src/+/3515012
[1]: https://chromium-review.googlesource.com/c/chromium/src/+/3526375

Bug: chromium:1300296
Change-Id: I912129d6c07634da3f4ccabf107c1b1b9b73ba1d