adb_profile_chrome_startup: add flags from fieldtrial_testing_config
Before starting the browser Telemetry takes the
fieldtrial_testing_config.json from chrome-src and sets the command line
arguments for the browser to assign fieldtrial values according to this
configuration.

In this change do the same for the `adb_profile_chrome_startup`. This
script is primarily used for running one-time local startup experiments
on Android. The change is good for consistency with the metrics recorded
by Telemetry (via Pinpoint, for example).

Bug: chromium:1302992
Change-Id: I5dba8f9e6985d80a27af15bee4bfb0cfcdd73b7e