Pinpoint: Implement A/B ordering randomization + no intermediate jobs between A and B.
Limitations:
- A or B going first is randomly determined per iteration,
  instead of trying to evenly distribute.

All of these will be addressed in a follow-up CL.
Given the approach in this CL is a sidestep vs. the status quo, that
this will allow us to flush out unforeseen complications with how jobs are
scheduled, and that we are still fleshing out the details on how device selection
interacts with A/B ordering, this is acceptable.

Change-Id: I105610399153904b6a88e1e5ee0d69516482819c