Pinpoint: Implement A/B ordering randomization + no intermediate jobs between A and B.
Limitations:
- A or B going first is randomly determined, instead of having half where
A go first, and the other half where B do.

All of these will be addressed in a follow-up CL.
Given the approach in this CL is a sidestep vs. the status quo, and that
this will allow us to flush out unforeseen complications with how jobs are
scheduled, this is acceptable.

Change-Id: I105610399153904b6a88e1e5ee0d69516482819c