[NetLog Viewer] Drop optional-chaining operator and upgrade Flask
I ran into a few issues while trying to follow the deployment
instructions linked from [0].

When I tried to run build_for_appengine.py, I saw that vulcanize was
failing when it hit the JavaScript optional-chaining operator (`?.`) in
dns_view.js. Apparently, vulcanize is deprecated and its last release
was five years ago in 2017. I tried out its successor, polymer-bundler,
but that suffered from a less verbose variant of the same problem; it
just silently failed to inline the dns_view.js script as well as the JS
file that defines `Polymer`, which caused a number of runtime failures
on load.

We should probably replace vulcanize with a new bundler, but I'm not
sure that polymer-bundler is a good idea. The Polymer library appears to
be deprecated as well, but maybe that doesn't include the CLI tool [1]?
Perhaps we should pick a new bundler entirely or just build something
ourselves with Golang's "text/template" library.

After generating the bundle, I tried to run the appengine dev server,
but that failed with a Python import error (see below). I noticed that
requirements.txt was asking for a very old version of Flask that was
released in 2013 [2]. I bumped the version to the latest release, and
the import error went away. I didn't investigate this one very
thoroughly.

    catapult/netlog_viewer/appengine % dev_appserver.py app.yaml
    [...]
    ImportError: cannot import name 'Markup' from 'jinja2' (/tmp/tmppe8EAp/lib/python3.9/site-packages/jinja2/__init__.py)

[0]: https://chromium.googlesource.com/catapult/+/master/netlog_viewer/CONTRIBUTING.md#deploying-to-netlog_viewer_appspot_com
[1]: https://polymer-library.polymer-project.org/
[2]: https://pypi.org/project/Flask/0.10.1/

Bug: chromium:1319431
Change-Id: Ic61533edfc9ee2ebaa14dffdb8192f64fa2eafcb