Make clock sync errors non-fatal
Makes failure to issue a clock sync marker during tracing shutdown
non-fatal. Ideally this would never happen, but ensuring that may not
be possible, and if it is, should probably be handled by someone more
familiar with tracing.

This only seems to occur after a failed test, so it seems reasonable to
shut tracing down more forcibly/less correctly in this case.

Bug: chromium:1320873
Change-Id: I6c5ae10b3e90dfaa818388163996605140eed9e8