Add --test-list option in typ arg_parser
This option is used as default in
testing/scripts/run_isolated_script_test.py in chromium. The script
also maintains a hardcoded list |KNOWN_TYP_TEST_RUNNERS| to identify
typ tests and then change that arg to --file-list.

This change adds --test-list as another name for --file-list option
at typ so the hardcoded list is not needed anymore.

Bug: 1322273
Change-Id: I4b7ae11a8b6baec169a19d46cef9c56ce29ff76b