Generalize symbol caching
Generalizes the minidump symbol caching used by desktop_browser_backend
to work on all platforms. This was previously omitted since desktop
platforms are the only ones that we can guarantee having a browser
binary on the host. The generalized version allows other backends to
signal that they don't have a binary and fall back to the old (no
caching) behavior.

This should result in caching symbols for all Chromium functional
testing since we always have a browser binary in those tests.

Bug: chromium:1323316
Change-Id: Iea1cacd0a77a096b303c801272453c793130a736