Add instant app support
These are applications that can be installed from a url
and as such run in a more restrictive environment.

This only requires an "instant" flag to adb.

This will allow us to test webview CTS tests as
instant apps.

ADB will throw an error if `--instant` is not a supported
option so we don't need to have special error handling
around if the SDK level is supported or not.

Bug: chromium:1082333
Change-Id: I2e1295e43f0b9c4637f558fa07fe2daa018810b3