android: Add force queryable flag to install
There are WebView CTS tests that need to use
a mock content provider.

Android 11 introduces new package visibilities
that require apps to declare these dependencies
in their manifest through a <queries> tag.

CTS unfortunately does not use this and instead
bypasses this with the flag --force-queryable
which makes an installed application visibile
to any application that calls it.

We'll have to use this as well to add these
additional tests.

Bug: chromium:868055
Change-Id: I5d29dd95dded75f2c0d3823beb8abaac86497c9a