[typ] Work around Mac file name limit
Adds a workaround to the artifacts implementation for Mac's 255
character for any one piece of a filepath. If a part of the path is
detected as being over the limit (likely a long test name), it gets
replaced with a SHA1 hash.

Also adds a few unittests for a similar Windows workaround that were
missing.

Bug: dawn:1446
Change-Id: I502ef1c6d879f12c1f1618aaaac294bf5105af90