[Trace Viewer] Remove the "Please file a bug with this trace" warning
This warning shows up quite often and causes a lot of bugs to be created
on the crbug.com by the users of the old trace viewer. Since the
development of the old trace viewer is frozen and the new trace viewer
(Perfetto UI) doesn't have this problem, it doesn't make sense to
continue to show this warning.

Bug: chromium:1300854
Change-Id: I3b87006e67616cd9415d436c1ef33623ef56e2e0