[fuchsia] Add cast-streaming-shell browser type
This adds support for a Cast Stremaing Shell browser in telemetry,
needed to run GPU integration tests using the Cast Streaming Receiver
feature on Fuchsia's WebEngine.
In addition, this also fixes the ids.txt file location for
WebEngineShell, following the code move from //fuchsia to
//fuchsia_web in Chromium.
Finally, This adds support for multiple ids.txt in the
symbolizer. This ensures we can get symbols for both WebEngine and
WebEngineShell / CastStreamingShell.

Bug: chromium:1330713
Change-Id: Ife2d4ff2abc5281bf393e927a0ee1f6d605f14eb