Remove _typ_runner usage
Removes usage of _typ_runner in SeriallyExecutedBrowserTestCase and
related files. Having a reference to the runner is contrary to typ's
intended design and breaks when running tests in parallel. Instead, the
child reference is made available earlier to achieve the same goals.

This does not remove _typ_runner itself yet, as there is a usage in
Chromium that will need to be removed first.

Bug: chromium:1336393
Change-Id: I006e36743659d41e235aaf5b284add94054cf6ec