Fix GetApplicationVersion for missing packages N+
GetApplicationVersion still returns a response on
Android from N onwards. We have to check against each
line and confirm if the package exists or not.

Test:
Ran this against Android emulators and confirmed that on Android M I saw no output. Confirmed that from Android N to S, I saw the Unable to find package message in adb.

Bug: chromium:1129166
Change-Id: I9331bbd02ed5ec24992a56bf0da35eecc176deff