[pinpoint] use gthread workers and increase timeout for pinpoint workers
When we send lots of requests to swarming, if the the request handling takes more than 30 seconds, we will see a worker timeout error:
https://pantheon.corp.google.com/logs/query;query=resource.type%3D%22gae_app%22%0Aresource.labels.module_id%3D%22pinpoint%22%0Aseverity%3D%2528DEFAULT%20OR%20DEBUG%20OR%20INFO%20OR%20NOTICE%20OR%20WARNING%20OR%20ERROR%20OR%20CRITICAL%20OR%20ALERT%20OR%20EMERGENCY%2529%0Aresource.labels.version_id%3D%220622-cleanup-9-py3%22%0A%22Requesting%20swarming%20task%20with%20parameters:%22;timeRange=2022-06-23T14:08:30.463Z%2F2022-06-23T14:15:30.463Z;cursorTimestamp=2022-06-23T14:09:50.977540Z?project=chromeperf-stage

The reason is:
when pinpoint requests scheduling swarming tasks, a handler will start to loop through the attempts and create swarming tasks for each of them. Based on the behavior in python 2 runtime, the loop may take as much as 50+ seconds, which exceeds the 30 second time out (https://docs.gunicorn.org/en/stable/settings.html#timeout).

Fix:
Use gthread as worker type should keep the worker alive. Also increase the timeout to match expectation.

Bug: chromium:1339080
Change-Id: Ia55426e85db7787bcccb97eba0557b9e790c8476