Remove frame_times from renderingMetric
The metrics in frame_times.html are noisy, we have switched reporting to
Graphics.Smoothness.PercentDroppedFrames2.AllSequences in:
https://chrome-internal-review.googlesource.com/c/infra/infra_internal/+/4835610

This change removes the non-CPU metrics from frame_times.html. The setup for
measuring CPU metrics has been moved to rendering_metric.hmtl.

We have also renamed frame_times.html to frame_segment_filter.html to be a
helper method. So that we can still filter for frame segments to measure the
CPU usage.

Bug: chromium:1135662
Change-Id: I2625a318675597a41428e658895ff3cca450a2a0