[typ] Fix TOC/TOU race in `Host.maybe_make_directory`
Previously, `Host.maybe_make_directory` checked if the target directory
did not exist before attempting to create it. It was possible for
another process to create the same directory and invalidate the
existence check. Example:
  https://chromium-swarm.appspot.com/task?id=5bbf788b214bc410&w=true

In this change, `maybe_make_directory` always attempts to create the
directory ("ask forgiveness, not permission").

Bug: chromium:1332629
Change-Id: I785fb20f57a72db9a4300b692b66560f5557e741