[typ] Plumb through number of jobs
Plumbs through the number of parallel jobs to children so that tests
can have access to that information. The specific use case in mind is
to dynamically increase test timeouts with higher job counts to account
for heavy tests slowing down a bit, but there may be additional cases
where this information will prove useful.

Bug: chromium:1336393
Change-Id: Id3bad55c13f9f5703a057e2b8841e7b07798461e