Update browser test parallel check
Updates the parallel check in SeriallyExecutedBrowserTestCase to be a
regular member method instead of a class method. The Chromium override
of this has already switched to this, and the one place where it is
actually called already uses it as a member method instead of a class
method. Ultimately, this allows suites to selectively run tests in
parallel/serial based on test name, etc. instead of being all or
nothing.

Submitted with No-Presubmit to work around unrelated pylint issues.

Bug: chromium:1336393
No-Presubmit: true
Change-Id: I26bf2b3963766f017b4ec793a5ffdbf94b0ac07b