Fix URL KeyError
Fixes issue we occasionally see where a KeyError is thrown when trying
to retrieve a JavaScript console message from the inspector by using
get() instead of assuming the presence of the key. Also preemptively
applies the fix to other keys in case it's possible for them to be
unset as well.

Change-Id: I54ebe31ce6be40a5d7b1d971f412704afd7f7670