systrace: Fix manually stopping tracine in Python3
Use six.moves.input() as a standin for raw_input() as it is
compatible with both python2 and 3. This is conveniant as Catapult
already includes this lib in third-party, and Systrace is already
making use of it.

The general recommended approach of using "from builtins import input"
to make python3 code compatible with python2 doesn't work in this case
unless we add a new third-party dependency, so using six is simpler.

Testing:
Test with both python2 and python3
- Run: python2 systrace/bin/systrace
- Run: python3 systrace/bin/systrace
verify both commands work.