Add apex support for Android
We would like to add support for apex files that
can be used to install mainline modules.

These files have been structured to be similar to
apks so that we can use the same tools.

A new API has still be introduced instead of using
the existing install method because we do not want
to do things like uninstall the APEX first, and we
need to perform a device reboot. The device reboot
needs to be done between each install which means
we should lock these installs to avoid race
conditions.

Test: This was also manually tested with an Android
emulator and an apex file.

Bug: chromium:134525
Change-Id: Ibbd63069001390053f029f75770563659de01edb