[pinpoint] logging get calls on api/isolate GET on pinpoint
When fixing the tests in handlers/isolate_tests.py, I found that webtest did not have a way to call Get with json data. On the handler, we have been expecting json data for both Get and Post. The Post calls are from recipe to update the isolate info when a build is finished, but I'm not sure where the Get calls are from.

From the log viewer:
https://pantheon.corp.google.com/logs/query;query=resource.type%3D%22gae_app%22%0Aresource.labels.module_id%3D%22pinpoint%22%0Aseverity%3D%2528DEFAULT%20OR%20DEBUG%20OR%20INFO%20OR%20NOTICE%20OR%20WARNING%20OR%20ERROR%20OR%20CRITICAL%20OR%20ALERT%20OR%20EMERGENCY%2529%0A%22api%2Fisolate%22%0AprotoPayload.method%3D%22GET%22;timeRange=P1D;cursorTimestamp=2022-07-28T19:05:18.264198Z?project=chromeperf
I did not see any get call on api/isolate.

Adding logs here to help figure out how Get calls are made.

Bug: chromium:1120195
Change-Id: Ib8a00f98032c1ebe25c8b837ec757e365bfbfffb