Add parent stack trace to thread reraiser
When the thread reraiser is used - for example
with Android adb commands - we only get the stack
trace of the reraiserthread when there is an exception.

This makes it more difficult to debug any issues,
especially if the issue happens on CI and is rare.

Adding logs of the parent stack trace so that
we can see where the exception originated from in our code.

Test: I induced an error into an adb method using this
Change-Id: I842988373f2f3965dea27350d294fc9a96d26715
and confirmed I saw logs.