[Catapult]: Use PersistentShell for device_util.RunShellCommand
This is a continuation of:
https://chromium-review.googlesource.com/c/catapult/+/1947706

"adb shell" is oddly expensive. It takes about 30ms on an emulator
and 100ms on a real device just to spin-up to do an echo.

Chaining commands from the test runner via multiple invocations
gets expensive fast. Currently chrome_public_test_apk makes 10
adb shell invocations per test which adds at least 300ms to each
test.

This CL changes device_util.RunShellCommand to spin up a
PersistentShell object which can be reused over the lifetime of
the script.

It also makes a fix to the PersistentShell for the "non-close"
mode to avoid a deadlock prone coding pattern that can occur if
simultaneously reading and writing from pipes to a popen()ed
subprocess.

Without any further changes, this takes test-to-test overhead
from ~650ms down to 250ms.

Bug: chromium:1021686
Change-Id: I0eaef44be9ce5475c5fe8aab11ddf5f7bcb91eab