[chromeperf] hosting multiple services using middleware
This CL will only affect those handlers in chromeperf (non pinpoint).

The current implementations uses utils.IsRunningFlask to tell whether we should run chromeperf using flask or webapp2. In fact, flask is python2/3 compatible. Once the handler is rewritten in flask, we should start running it in production.

werkzeug provides a way to allow multiple apps to handle requests from different paths. It has the limitation that the request redirect is based on a mapping from path to app. The current pinpoint is not designed to separate the handlers by url path. Thus, we need to create multiple instances of flask and use one to handle each single handler.

It will cause resource issue which we don't know how much it is yet. Though, it allows running flask handlers after it is checked in. The previous implementation flips all handlers from webapp2 to flask, which could be more risky.

Change-Id: I248c6360805770626817d86a12374bfc9db42be9