Make async tracks with < 30 subrows default expanded
A internal profilter based on trace viewer uses streaming data loading for their trace viewer, which resets
the async slice group expanded state to the default state on every model
update (which can happen when the user moves around the trace or zooms
in/out). It's difficult to retain the expanded/collapsed state of the
slice group tracks across these updates, so as a usability improvement,
set the default state to be expanded when there are less than 30
subrows.

Note that for non-async tracks, the default is expanded for tracks with
less than 100 subrows:
https://chromium.googlesource.com/catapult/+/refs/heads/main/tracing/tracing/ui/tracks/slice_group_track.html#32

Change-Id: Ie415e26af4bd1147fb39bf2f85fe8ce37215f69b