make async slice group track expand by default
Background:
The slice_group_track component is expanded by default, but the async component is not.
This will make it hard for user to browse the data when the trace viewer data/model got updated (eg. when user expand the async track and expore the data, as user zoom in/out, our incremental loading mechanism will refresh the data, which trigger the UI to initial state - the async tracks that the user are browsing will be collapsed, breaking the previous view that the user was browsing at.)

ref for implementations on sync slice group:
https://chromium.googlesource.com/catapult/+/refs/heads/main/tracing/tracing/ui/tracks/slice_group_track.html#32

Change-Id: Ie415e26af4bd1147fb39bf2f85fe8ce37215f69b