[Catapult]: Create PersistentShell for adb.
This CL was reverted with
https://chromium-review.googlesource.com/c/catapult/+/3929220.

This CL is the same as the revert, but avoids checking if the device is
ready when a persistent shell isn't desired as some builders (mac-rel, linux-rel) call a test that uses an adb_wrapper but have not devices.

This CL was synced adn tested on this change.
https://chromium-review.googlesource.com/c/chromium/src/+/3823153

Bug: chromium:1021686
Change-Id: I909d81ecc3301419e2a1230a56098537c990ea34