[gpu][telemetry] Use the CFv2 version of web_engine_shell
Changes the gpu/telemetry tests to use the new CFv2 version of
web_engine_shell. This is accomplished by launching web_engine_shell
with `ffx test run`. This results in the remote debugging port being
printed to stdout, as ffx combines the stdout and stderr streams from
the launched component.

Bug: chromium:1352062
Change-Id: I16e63c148b7dd96c8f5e58ff2cb8cbfe1bbb2311