[chromeperf] Reuse the xsrf logic from oauth2client
Reusing the logic in oauth2client.xsrfutil. We are removing the usage of oauth2client, as it has been deprecated long ago. Though, the xsrf related logic is still valid and I did not find a handy replacement. As the logic itself is trivial, I copied and updated it so that to our current behaviour consistent.

Please let me know if any regulation I should be aware of.

Bug: chromium:1363763
Change-Id: I1b96a57776fd8569b410b8369d1f5e7cbd615560