Devil: suppress errors in Logcat
The logcat wrapper methods currently use 'check_error=False', however
this is not respected and an Exception is still thrown when a logcat
command fails. This aims to fix this with 3 changes:

* Propagate the 'check_error' parameter through the call stack
* Change the meaning of 'check_error' for adb._RunAdbCmd(). This method
  used to throw errors if `status != 0`, even if the caller specified
  `check_error=False`. Now callers must specify `check_error=True` if
  they want this method to throw errors as Exceptions.
* Update adb._Shell() to use `check_error=True` (the default value)
  since this is one of the call sites which expected Exceptions.

The only other call site which uses `check_error=False` is the Logcat()
function, which is where we want this behavior change.

Bug: chromium:1378135
Test: run_webview_instrumentation_test_apk -f AnyTest.*
Test: ./devil/bin/run_py3_tests
Change-Id: If9f6f1ce469e19653cc51c99fe5544f67e19b84d