Add default retries and timeout for WaitUntilFullyBooted
This CL changes WaitUntilFullyBooted's decorator to
WithTimeoutAndConditionalRetries and adds a default timeout of
60 seconds and 2 retries by default. WaitUntilFullyBooted will
only retry if any of the adb command's it runs times out.

This CL also modifies the WithTimeoutAndConditionalRetries decorator
so that it gets the default retry count and timeout from the default
arguments of the decorated function.

Bug:chromium:1286468
Change-Id: I3eeaa47428c1aaf0b0463c59308c9b77e1413d08