[tbmv2] Fix isTelemetryInternalRanges
The previous implementation did not work when event markers were in
different processes, and it did not take event name into account when
pairing up start and end events. See associated bug for more details.

This CL changes the concept of telemetry internal events from process
scoped to global/model scoped, since we should not be looking at events
from other processes anyways during these "internal" time ranges. This
became a bug while implementing the new coreWebVitals lab metric, which
is derived exclusively from events in the browser process, so the metric
was reporting values from the internal ranges.

This will affect metrics of any benchmark that uses
cacheTemperature. See this search for an incomplete list:
https://source.chromium.org/search?q=cache_temperature_module%20f:py%20-f:ct&ss=chromium

However, the metrics will be more correct now instead of being
silently misleading. Hopefully autotriage will find this CL as the
culprit for the regressions and we can wontfix the bugs.

Bug: chromium:1381260
Change-Id: I9c39ec71f09ab888a30a7573ae532f16b21bd6a9