Disallow non-root startup tracing
Disallows the use of startup tracing on non-rooted Android devices. The
current hard-coded path (which is in both Telemetry and Chrome) is not
accessible without root permissions, so file creation fails. It may be
possible to change the path to a non-root-accessible one, but initial
attempts to do so failed due to Chrome not being able to read the file.

Since the only current use case for tracing on non-rooted devices is
GPU testing and those tests don't use startup tracing, we can skip
trying to add support for now.

Bug: chromium:1358240
Change-Id: I7afa11b57f7f86eb0eecb7aecdf9685529a393fe