[chromeperf] enable add_point and add_histograms in flask
The flask handler has successfully received data from system_health.common_desktop on linux-perf-fyi:
https://chromeperf.appspot.com/report?sid=e60508ac65c01b5154fb61008338bdaa24f992525b32b671551c174e588389bb

And we can see the logs are from flask and under 'upload' service:
https://pantheon.corp.google.com/logs/query;cursorTimestamp=2022-11-23T06:53:54.547078Z;query=resource.type%3D%22gae_app%22%0Aresource.labels.module_id%3D%22upload%22%0A%22add_histograms_flask%22%0Atimestamp%3D%222022-11-23T06:53:54.547078Z%22%0AinsertId%3D%22637dc3830000527d413be078%22;timeRange=PT1H?project=chromeperf

Though, I don't have a easy way to test add_point. I will start to assign small amount of the traffic to this check-in for testing.

Bug: chromium:11298177
Change-Id: I091991bba22089daf27fcfd17fb7aa318b28a3d8