[fuchsia][perf] Read more lines at a time from stderr.
The current method for reading lines from STDERR of the browser
process assumes that the DevTools port will be available within the
first 6 reads or so. This will allow more output to be read at once,
removing this limitation, and truly making it a 60s limit instead of
a 6-line/60s limit.

Change-Id: Ie4fa276323ea60b4ede8335743e0884580873258