Use fastboot command `flashall` to flash the device.
So after fastbootd is introduced [1], the flash process gets more complicated,
as we need to reboot the devices into bootloader and fastbootd at different
flash phrases, and more image types are added to the flash process.

Instead of updating the flash_device.py, we can simply rely on the fastboot
native command `fashboot flashall`[2] to do the job.

[1] https://source.android.com/docs/core/architecture/bootloader/fastbootd#fastboot-and-bootloader
[2] https://android.googlesource.com/platform/system/core/+/master/fastboot/fastboot.cpp

Bug: chromium:1379427
Change-Id: I1a487c962120fcdbdde7bdba6674e82c71c7937c