[typ] Add rdb output file support
Adds support for --rdb-content-output-file, which causes typ to write
data that would be sent to ResultSink to the specified file instead.
This is currently only intended for use in a Skylab environment, as
their setup currently prevents tests from using native ResultSink
integration.

The resulting file contents are a JSON list of results, which is
suitable to be later passed to ResultSink as-is as testResults key's
value.

Bug: chromium:1233678
Change-Id: Ic3d21f863a0ef7e1a43711c14c5d7d0ec1ec0527