[chromeperf] move forward versions of chromeperf
Due to the datahook issue, I have manually moved all traffic to the version with the fix. As the version change has been in prod for a week, I will call it stable this time. Otherwise we need to pick a really old version to avoid the datahook bug.

This CL will set traffic to 'upload' service to py3 versions.

Also, setting the traffic splitting rule to random for upload and upload-processing.

Change-Id: I665d6c326beecf18341bb209507cecc6e6e94aee