[chromeperf] Update proto files to allow correct absolute imports in python 3
The generated _pb2.py files have import issue in Python 3. E.g., in sheriff_config_pb2.py, the following statement does not work:
  import sheriff_pb2.py
Instead, need something like:
  from dashboard import sheriff_pb2

As suggested from proto team, we need to run the 'protoc' one level up so that the 'dashboard' can be part of the import statement in the _pb2.py. Thus, we do the following in this CL:
 - move the Makefile (which runs the 'protoc') one level up
 - The generated py files will stay in proto/ folder.
 - add __init__.py in the proto/ folder to avoid import issue in unit tests
 - update the other files to accommodate this change.

Bug: chromium:1241093
Change-Id: Ic6490824904d190ebd5d326997e63577c6a7789d