Telemetry client should not set origin
The origin header is a forbidden header name[1] which means that generally it should not be up to the user to decide upon it. The origin header might be used by the server to enforce cross-origin policies and
non-browser clients should not need it in general.

I realise that this is a third party library but I see that it was not updated for 5 years and there were changes made to it. So is it a work at this point? The newer version comes with an option to suppress sending the origin header so we can either remove it in this version
or use the option if the library can be updated.

[1]: https://developer.mozilla.org/en-US/docs/Glossary/Forbidden_header_name

Bug: 1385982
Change-Id: I591392e2cb0de1e722f98e3611b765be47667c8e