[chromeperf] fix path info loading in python 3
The pre get hook tried to load the path info from request.path and failed, because there no flask context when the request is from defer queue. As a result, an extra filter was added to filter out the internal-only records. We need to change to way to find the path info.

Also some changes:
 - refresh_jobs_test failed because it tries to check for flask global. Noticed that '/_ah/queue/deferred' is not in PATH_INFO, added using PatchEnviron.
 - Update the usage of http to avoid thread safety issue, as we did in https://chromium-review.googlesource.com/c/catapult/+/3910640


Bug: chromium:1298177
Change-Id: Ie0374ccd3e4a9ea5e30761c0b335d6d0f769e45a