[chromeperf]: Add a group membership check for test migration tool.
Currently the tool requires access to the admin group (which potentially gives a user more than required access). A new group has been set up to cater to only this tool and the idea is users can request temporary access to this group to access the tool and is removed once the migration is done.

Change-Id: Ia21ede21b5aee70cb1cbff1a9f1482e59f0130c2