Support both path types with traling slash and without it for --chromium-output-dir
Path from --chromium-output-dir path is forwarded to build_path in
desktop_browser_finder.py. If this path ends with trailing slash then
desktop browser finder breaks due to the behavior of os.path.basename:
for '/foo/bar' it returns 'bar', but for '/foo/bar/' it returns ''.
You can find more information in official docs:
https://docs.python.org/3/library/os.path.html#os.path.basename

This CL is designed to fix this problem by deletion of path separator
from the end of path if it exsists.

Change-Id: I1f0bd74dddf594fc59cb019d9e2164ebf6b84663