[remote-linux] Add `interface/has_interface` to platform_backends.
`interface` is intended to be a property to connect to a remote
or local interface of a device.
This is similar to the `cri` property of the cros_platform_backend,
but extended to other platforms. If an interface is present, it is
used for the platform instead of the default (which is a subprocess
call, which is the same if the interface is local).

This is the same as `device` on android_platform_backend, but the
`device` interface of Android differs from the linux/cros interfaces.

Change-Id: I838668a188fd5aefe178737125567edc427d149f