Modify _ProcessTest to use max_window_size from subscriptions instead of overridden_anomaly_config property.
The problem is that _ProcessTest uses max_window_size from overridden_anomaly_config property to fetch rows from the test. It disregards subscriptions' max_window_size. It assumes test and max_window_size are one-to-one.

This CL refactors _ProcessTest and _ProcessTestStat to use the appropriate max_window_size for each of the test's subscriptions.

Bug: chromium:1410304
Change-Id: I21a49411b45a8ba2c1c1637604c40385e85c30ac