Add stack summaries
Adds stack summaries to the error raised when unsymbolized minidumps are
left over by a test. A stack summary is the N function calls leading up
to the crash, where N is currently 3.

This adds some benefit for humans, as it will be easier to tell at a
glance what the crash cause is. However, this is largely intended to
improve LUCI Analysis' ability to cluster relevant failures together.
Currently, all cases of tests leaving unsymbolized minidumps behind are
clustered together, which is not very useful since there are many
different root causes. Including the stack summary should cause failures
to only be grouped if they appear to have the same root cause.

Bug: chromium:1412508
Change-Id: I2cc62632002281da04778dd028639eeec10c548b