Add A23-specific workaround
Adds a profile directory and profile removal workaround for the Samsung
A23, which will allow it to use a code path closer to what rooted
devices use. This is possible because of the filesystem that this
particular device uses.

This is written in a device-agnostic way so that future non-rooted
devices can easily add similar workarounds in the future if necessary.

Bug: chromium:1383609
Change-Id: I81e48aaf33e562e35dcb0e6332816a8833169f76