Fix the frozen job refresh.
The original frozen job refresh code run deferred.defer to push the _ProcessFrozenJob operation to the deferred queue, which is the "default" queue under the "default" service.

Then _ProcessFrozenJob and job._Schedule() will be run from the "default" service.
Then the following code will be run from the "default" service:
      task = taskqueue.add(
          queue_name='job-queue',
          url='/api/run/' + self.job_id,
          name=task_name,
          countdown=countdown)

Yet the /api/run/job_id api is only available in the "pinpoint" service, and the job-queue won't able to push the task to the /api/run/job_id api under the "default" service.

So, the frozen job refresh doesn't work.

This fix is to change the _ProcessFrozenJob from deferred queue to be run locally from pinpoint service.

Change-Id: I4f38389d630ca310756abf3cd4090a7b971c60c8