Reduce the auto bisect speculation level from 2 to 1 to speed up auto bisect jobs
Speculation level = 2 means whenever need to add addition changes it always try to add 3 new changes, it was suppose to be useful when attempting to bisect a potentially large range of revisions
quickly when finding one or more culprit changes for performance regressions. It was based on the assumption that there is unlimited testing bots to run the tests.
But the reality is that we only have limited testing bots, many of the testing tasks pending in Luci Swarming longer than its actual test run time.

This change will reduce the number of Swarming tasks, and consequently reduce the pending time of the Swarming tasks. While it won’t affect the accuracy of auto bisection jobs.

For example:
level = 2 will add N1/4, N1/2, N3/4 for the changes N1/8 & N7/8
                 N1/2
       N1/4                N3/4
  N1/8                          N7/8

level = 1 will add N1/2 for the changes N1/8 & N7/8
                 N1/2
  N1/8                          N7/8



Change-Id: Ibb0b7b7d39c9dc3172e05fac0beae4bd99e5e1af