Support Windows minidump annotation extraction
Adds support for extracting annotations from a minidump on Windows. This
was previously added for all platforms, but it turns out that the
required minidump_dump binary does not exist for Windows. Instead, we
now use the dump_minidump_annotations binary on Windows.

Also drive-by fixes values that contain spaces not being extracted
properly.

Bug: chromium:1006331
Change-Id: I6ebb0dd2222462758b784009ea19300065dcdd99