Temporarily remove Fuchsia proxy
Temporarily removes the --proxy-server argument from the list of Chrome
startup args on Fuchsia. This argument currently does nothing since it
is not supported, but we cannot add support on Chrome's end while this
arg is being passed in AND the existing forwarder-based workaround is
still in place in Telemetry. Once this lands, support can be added on
Chrome's end, after which this arg can be re-added and the workaround
removed.

Other relevant CLs that must land after:
* Chrome argument support crrev.com/c/4298514
* Use of argument in Telemetry crrev.com/c/4298549

Bug: chromium:1354797
Change-Id: I5aa3f52d6187ec023121a93961ed404f5d6a7128