[devil] Special case screenshot logic
Adds an alternative code path for capturing screenshots on certain
devices and enables it on Pixel 4 and Pixel 6 devices.

The affected devices often hang when running
"screencap -p path/to/file.png", but for whatever reason, work fine
with "screencap -p > path/to/file.png", likely due to some bug in
the tool when creating/opening the specified file.

Bug: chromium:1446736
Change-Id: I0fce2e8ed0dbe5257fd8137eec1eb8cbe2886569