[chromeperf] Add configuration for the projects which are migrated from Monorail to Buganizer
- Added a configuration named buganizer_projects to have a list of projects which have been migrated from Monorail to Buganizer. If the project is listed, the perf issues should be filed to Buganizer.
- Added a parameter for the existing GET handler of edit_site_config. It will be called by the perf_issue_service to decide where to file the issue.

The pros of this change is: we can have a UI to update the config easily. The cons is: need more changes on code. Other then this CL, we need a client to call from perf_issue_service to chromeperf to read the config.

An alternative is: to create a temparary entiry direct from the datastore and let perf_issue_service to get the value directly. We need to manually update the config from datastore.

Bug: chromium:1447930
Change-Id: I2d20020bef6073feb72b4f5e88efa51ffce6567a